[JIT] Specialize autograd zero: actually remove the original graph after we created its versioned copy.
The original code assumed that the versioning if was inserted in the
beginning of the graph while in fact it was inserted in the end. We're
now also not removing `profile_optional` nodes and rely on DCE to clean
it up later (the reason we're not doing it is that deletion could
invalidate the insertion point being used).

ghstack-source-id: a6be26cdf49ced5ed92c596fab0485cbb4cac890
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43900