[JIT] Improve error message for unsupported Optional types
**Summary**
This commit improves the error message that is printed when an
`Optional` type annotation with an unsupported contained type is
encountered. At present, the `Optional` is printed as-is, and
`Optional[T]` is syntatic sugar for `Union[T, None]`, so that is what
shows up in the error message and can be confusing. This commit modifies
the error message so that it prints `T` instead of `Union[T, None]`.

**Test Plan**
Continuous integration.

**Fixes**
This commit fixes #42859.

ghstack-source-id: 630ca11b2a591c54f8cfca391a8b6c1456bf8a62
Pull Request resolved: https://github.com/pytorch/pytorch/pull/44054