[JIT] Remove profile nodes before BatchMM.
Currently we're removing prim::profile nodes and embed the type info
directly in the IR right before the fuser, because it is difficult to
fuse in a presence of prim::profile nodes. It turns out that BatchMM has
a similar problem: it doesn't work when there are prim::profile nodes in
the graph. These two passes run next to each other, so we could simply
remove prim::profile nodes slightly earlier: before the BatchMM pass.

ghstack-source-id: fe79a08e6a113e7cb391317fe0c0e19eeca8e758
Pull Request resolved: https://github.com/pytorch/pytorch/pull/43961